replace int type with byte

+++ memory-optimization/src/main/java/org/example/App.java
@@ -4,14 +4,14 @@ public class App {
     public static void main(String[] args) {
         long nanoTimeBefore = System.nanoTime();
         System.out.println(nanoTimeBefore);
-        int[][][] AMD = new int[100][100][100];
-            int res = 0;
+        byte[][][] AMD = new byte[100][100][100];
+            byte res = 0;
 
-            for (int i = 0; i < 100; i++)
+            for (byte i = 0; i < 100; i++)
             {
-                for (int j = 0; j < 100; j++)
+                for (byte j = 0; j < 100; j++)
                 {
-                    for (int k = 0; k < 100; k++)
+                    for (byte k = 0; k < 100; k++)
                     {
                         AMD [i] [j] [k]++;
                     }
